# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestió de Magatzem",
        "description": "Documentació Swagger del projecte final",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/productes": {
            "get": {
                "tags": [
                    "Productes"
                ],
                "summary": "Llistar tots els productes",
                "operationId": "e8ab39ee27e015af7baca5a9e6bb653c",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Productes"
                ],
                "summary": "Crear un producte",
                "operationId": "734ab08ee920bc83a59c6287856c2ba6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "referencia": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "descripcio": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "quantitat": {
                                        "type": "integer"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producte creat"
                    },
                    "422": {
                        "description": "Validació fallida"
                    }
                }
            }
        },
        "/api/productes/{id}": {
            "get": {
                "tags": [
                    "Productes"
                ],
                "summary": "Obtenir un producte per ID",
                "operationId": "f736d1d164c13ea102af7e64182ebe3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producte trobat"
                    },
                    "404": {
                        "description": "Producte no trobat"
                    }
                }
            },
            "put": {
                "tags": [
                    "Productes"
                ],
                "summary": "Actualitzar un producte",
                "operationId": "dd3af72f5853c41ff299519e322e6c1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "referencia": {
                                        "type": "string",
                                        "maxLength": 100
                                    },
                                    "descripcio": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "quantitat": {
                                        "type": "integer"
                                    },
                                    "categoria_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producte actualitzat"
                    },
                    "404": {
                        "description": "Producte no trobat"
                    },
                    "422": {
                        "description": "Validació fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Productes"
                ],
                "summary": "Eliminar un producte",
                "operationId": "4b34a3283287787c634dcdbc73a6f33d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producte eliminat"
                    },
                    "404": {
                        "description": "Producte no trobat"
                    }
                }
            }
        },
...
    },
...
}